% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercom_main_coexp.R
\docType{data}
\name{Background_signaling_interactome}
\alias{Background_signaling_interactome}
\title{Background signaling interactome}
\description{
Background signaling interactome
}
\author{
Sascha Jung \email{sjung@cicbiogune.es}
}
\keyword{data}
