% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercom_library_coexp.R
\name{bool.data}
\alias{bool.data}
\title{Booleanize gene expression data}
\usage{
bool.data(exp.tbl, expr.thrs = 0)
}
\arguments{
\item{exp.tbl}{Gene expression dataframe.}

\item{expr.thrs}{Expression threshold.}
}
\value{
Gene expression data frame with booleanized expression.
}
\description{
The function booleanizes the expression matrix based on expression threshold.
It also removes genes not expression in any cell.
}
