% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercom_library_coexp.R
\name{comp_score_tf}
\alias{comp_score_tf}
\title{Calculating compatibility scores}
\usage{
comp_score_tf(t, s, g)
}
\arguments{
\item{t}{A character vector of target node names present in graph g}

\item{s}{A vector of source node names present in graph g}

\item{g}{An igraph graph object}
}
\value{
Vector of compatibility scores for each source node s
}
\description{
Calculates the compatibility between a list of source nodes and a list of target nodes depending
on their expression values
}
