% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainlib.R
\name{compatability_score}
\alias{compatability_score}
\title{Calculating activation probabilities}
\usage{
compatability_score(x, y, int)
}
\arguments{
\item{x}{A list of compatibility scores}

\item{y}{The steady state distribution of a Markov chain}

\item{int}{Character vector of intermediate signaling molecules to consider}
}
\value{
Data frame of activation probabilities per gene
}
\description{
Calculates the probabilities that signaling molecules are active
}
