% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainlib.R
\name{Data_preprocessing}
\alias{Data_preprocessing}
\title{Input Data to Markov Chain}
\usage{
Data_preprocessing(input_data, cutoff, species = NULL)
}
\arguments{
\item{input_data}{Data frame of input data. Rows correspond to genes, columns to cells}

\item{cutoff}{Maximum number of zero-value genes, above this cutoff the genes are excluded}

\item{species}{The species from which the input data has been obtained (UNUSED)}
}
\value{
Markov chain of the signaling interactome
}
\description{
The function computes a Markov chain of the signaling interactome.
}
