% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainlib.R
\name{gene.coexp}
\alias{gene.coexp}
\title{Compute Co-expression Between Genes}
\usage{
gene.coexp(exp.tbl, gene.frame, ncores = 4)
}
\arguments{
\item{exp.tbl}{Gene expression matrix/data frame to calculate coexpression of genes.}

\item{gene.frame}{A data frame with genes. The coexpression is calculated for genes in each row.}

\item{ncores}{Number of cores to be used (Default : 4)}
}
\value{
gene.frame data frame with an additional column corresponding to the number of cells coexpressing the genes.
}
\description{
The function calculates the coexpression of genes in each row of a given data frame using the gene expression table.
}
