% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercom_library_coexp.R
\name{get.cons.tfs}
\alias{get.cons.tfs}
\title{Get conserved TFs}
\usage{
get.cons.tfs(exp.tbl, quantile = 0.95)
}
\arguments{
\item{exp.tbl}{Gene expression dataframe.}

\item{quantile}{Minimum fraction of cells expressing the Tf (default: 0.95)}
}
\value{
List containing all TFs and data frame including only TFs expressed in the required fraction of cells
}
\description{
The function computes TFs expressed in a given percentage of cells.
}
