% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HUMAN_dummy.var}
\alias{HUMAN_dummy.var}
\title{Dummy variable (Human)}
\format{
A character with the content "DUMMY"
}
\usage{
HUMAN_dummy.var
}
\description{
Dummy variable (Human)
}
\details{
This variable is used to represent the niche in the Markov Chain model of intracellular
signaling, such that the chain becomes ergodic.
}
\author{
Sascha Jung \email{sjung@cicbiogune.es}
}
\keyword{data}
