% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HUMAN_Ligands}
\alias{HUMAN_Ligands}
\title{Secreted ligands (Human)}
\format{
A character vector containing secreted ligands
}
\usage{
HUMAN_Ligands
}
\description{
Secreted ligands (Human)
}
\details{
Secreted ligands have been identified based on their Uniprot annotation
(c.f. ./data-raw/process.R). Only ligands being annotated as "Secreted" were considered.
}
\author{
Sascha Jung \email{sjung@cicbiogune.es}
}
\keyword{data}
