% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HUMAN_LR}
\alias{HUMAN_LR}
\title{Ligand-receptor interactions (Human)}
\format{
A data frame with 1450 rows and 3 variables:
\describe{
\item{\code{L_R}}{character Ligand receptor interactions in the form "Ligand_Receptor"}
\item{\code{Ligand}}{character Gene symbol of the secreted ligand}
\item{\code{Receptor}}{character Gene symbol of the receptor}
}
}
\usage{
HUMAN_LR
}
\description{
Ligand-receptor interactions (Human)
}
\author{
Sascha Jung \email{sjung@cicbiogune.es}
}
\keyword{data}
