% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HUMAN_Receptors}
\alias{HUMAN_Receptors}
\title{Receptors (Human)}
\format{
A character vector containing human receptors
}
\usage{
HUMAN_Receptors
}
\description{
Receptors (Human)
}
\details{
Human receptors
}
\author{
Sascha Jung \email{sjung@cicbiogune.es}
}
\keyword{data}
