% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HUMAN_TF_TF_interactions}
\alias{HUMAN_TF_TF_interactions}
\title{TF-TF-interactions (Human)}
\format{
A data frame with 12203 rows and 3 variables:
\describe{
\item{\code{Source}}{character Gene Symbol of the regulating transcription factor}
\item{\code{Target}}{character Gene Symbol of the regulated transcription factor}
\item{\code{Effect}}{integer Whether the effect is activating - denoted as '1' - or inhibiting - denoted as '-1'}
}
}
\usage{
HUMAN_TF_TF_interactions
}
\description{
TF-TF-interactions (Human)
}
\author{
Sascha Jung \email{sjung@cicbiogune.es}
}
\keyword{data}
