% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HUMAN_tfs}
\alias{HUMAN_tfs}
\title{Transcription factors (Human)}
\format{
Character vector of length 1665.
}
\usage{
HUMAN_tfs
}
\description{
Transcription factors (Human)
}
\details{
The vector of human TFs corresponds to the values in the "Symbol" column of the
HUMAN_tf.db dataset
}
\author{
Sascha Jung \email{sjung@cicbiogune.es}
}
\keyword{data}
