% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainlib.R
\name{integrate_sig_TF}
\alias{integrate_sig_TF}
\title{Integrate Signaling and Transcriptonal Regulation}
\usage{
integrate_sig_TF(g, x, deg, non_interface_TFs, TF_TF_interactions)
}
\arguments{
\item{g}{Markov chain of a signaling interactome}

\item{x}{Steady state distribution of the Markov chain underlying graph g}

\item{deg}{Data frame of differentially expressed genes (one column only)}

\item{non_interface_TFs}{TFs that are not regulated by any active signaling pathway}

\item{TF_TF_interactions}{Regulatory interactions between transcription factors}
}
\value{
Markov chain of the signaling interactome including downstream TF targets
}
\description{
The function integrates downstream TF targets in the signaling network
}
