% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercom_main_coexp.R
\docType{data}
\name{intermediates}
\alias{intermediates}
\title{Intermediate signaling molecules}
\description{
Intermediate signaling molecules
}
\author{
Sascha Jung \email{sjung@cicbiogune.es}
}
\keyword{data}
