% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercom_library_coexp.R
\name{Markov_chain_stationary_distribution}
\alias{Markov_chain_stationary_distribution}
\title{Obtain Stationary Distribution of Markov Chain}
\usage{
Markov_chain_stationary_distribution(subg)
}
\arguments{
\item{subg}{An igraph object representing a Markov Chain. Must be ergodic.}
}
\value{
Stationary distribution of Markov chain
}
\description{
The function computes the stationary distribution of a given Markov chain
}
