% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOUSE_Background_signaling_interactome}
\alias{MOUSE_Background_signaling_interactome}
\title{Intracellular signaling interactions of mouse genes}
\format{
A data frame with 106175 rows and 3 variables:
\describe{
\item{\code{Source}}{character Gene symbol of the regulating gene}
\item{\code{Target}}{character Gene symbol of the regulated gene}
\item{\code{Effect}}{integer Whether the effect is activating - denoted as '1' - or inhibiting - denoted as '-1'}
}
}
\usage{
MOUSE_Background_signaling_interactome
}
\description{
Intracellular signaling interactions of mouse genes
}
\details{
This dataset is used for creating a Markov Chain model of intracellular signaling
to identify significantly active paths from receptors to downstream TFs.
}
\author{
Sascha Jung \email{sjung@cicbiogune.es}
}
\keyword{data}
