% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOUSE_dummy.var}
\alias{MOUSE_dummy.var}
\title{Dummy variable (Mouse)}
\format{
A character with the content "Dummy"
}
\usage{
MOUSE_dummy.var
}
\description{
Dummy variable (Mouse)
}
\details{
This variable is used to represent the niche in the Markov Chain model of intracellular
signaling, such that the chain becomes ergodic.
}
\author{
Sascha Jung \email{sjung@cicbiogune.es}
}
\keyword{data}
