% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOUSE_intermediates}
\alias{MOUSE_intermediates}
\title{Intermediate signaling molecules (Mouse)}
\format{
A data frame with 2493 rows and 1 variables:
\describe{
\item{\code{Gene}}{character Gene symbols of intermediate signaling molecules}
}
}
\usage{
MOUSE_intermediates
}
\description{
Intermediate signaling molecules (Mouse)
}
\details{
Intermediate signaling molecules are used within the Markov Chain model of
intracellular signaling and constitute all genes that are not at the interface to transcriptional
regulation. Importantly, they also contain receptors.
}
\author{
Sascha Jung \email{sjung@cicbiogune.es}
}
\keyword{data}
