% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOUSE_Ligands}
\alias{MOUSE_Ligands}
\title{Secreted ligands (Mouse)}
\format{
A character vector containing secreted ligands
}
\usage{
MOUSE_Ligands
}
\description{
Secreted ligands (Mouse)
}
\details{
Secreted mouse ligands have been mapped from secreted human ligands using the
homologene R-package (c.f. ./data-raw/process.R).
}
\author{
Sascha Jung \email{sjung@cicbiogune.es}
}
\keyword{data}
