% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOUSE_LR}
\alias{MOUSE_LR}
\title{Ligand-receptor interactions (Mouse)}
\format{
A data frame with 1422 rows and 3 variables:
\describe{
\item{\code{L_R}}{character Ligand receptor interactions in the form "Ligand_Receptor"}
\item{\code{Ligand}}{character Gene symbol of the secreted ligand}
\item{\code{Receptor}}{character Gene symbol of the receptor}
}
}
\usage{
MOUSE_LR
}
\description{
Ligand-receptor interactions (Mouse)
}
\details{
Ligand-receptor interactions have been obtained by obtaining mouse orthologs of
human ligand-receptor interactions using the "homologene" R-package (c.f. "./data-raw/process.R").
}
\author{
Sascha Jung \email{sjung@cicbiogune.es}
}
\keyword{data}
