% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOUSE_non_interface_TFs}
\alias{MOUSE_non_interface_TFs}
\title{Non-interface TFs (Mouse)}
\format{
A data frame with 323 rows and 1 variables:
\describe{
\item{\code{Gene}}{character Gene symbols of non-interface TFs}
}
}
\usage{
MOUSE_non_interface_TFs
}
\description{
Non-interface TFs (Mouse)
}
\details{
Non-interface TFs are those transcription factors that are not direct targets of
signaling interactions.
}
\author{
Sascha Jung \email{sjung@cicbiogune.es}
}
\keyword{data}
