% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOUSE_Receptors}
\alias{MOUSE_Receptors}
\title{Receptors (Mouse)}
\format{
A character vector containing receptors
}
\usage{
MOUSE_Receptors
}
\description{
Receptors (Mouse)
}
\details{
Mouse receptors have been mapped from human receptors using the
homologene R-package (c.f. ./data-raw/process.R).
}
\author{
Sascha Jung \email{sjung@cicbiogune.es}
}
\keyword{data}
