% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOUSE_tf.db}
\alias{MOUSE_tf.db}
\title{Database of mouse TFs}
\format{
A data frame with 1636 rows and 6 variables:
\describe{
\item{\code{Species}}{character The species this TF belongs to. Only contains one value ("Mus_musculus")}
\item{\code{Symbol}}{character The gene symbol of the TF}
\item{\code{Ensembl}}{character The ensembl gene id of the TF}
\item{\code{Family}}{character The protein family the TF belongs to}
\item{\code{Protein}}{character The ensembl protein id of the TF}
\item{\code{Entrez.ID}}{character The entrez gene id of the TF}
}
}
\source{
Animal TF DB
}
\usage{
MOUSE_tf.db
}
\description{
Database of mouse TFs
}
\details{
Dataset of all mouse TFs catalogued in the Animal TF DB.
}
\author{
Sascha Jung \email{sjung@cicbiogune.es}
}
\keyword{data}
