% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOUSE_TF_TF_interactions}
\alias{MOUSE_TF_TF_interactions}
\title{TF-TF-interactions (Mouse)}
\format{
A data frame with 10239 rows and 3 variables:
\describe{
\item{\code{Source}}{character Gene Symbol of the regulating transcription factor }
\item{\code{Target}}{character Gene Symbol of the regulated transcription factor}
\item{\code{Effect}}{integer Whether the effect is activating - denoted as '1' - or inhibiting - denoted as '-1'}
}
}
\usage{
MOUSE_TF_TF_interactions
}
\description{
TF-TF-interactions (Mouse)
}
\details{
Transcriptional regulatory interactions between TFs. Mouse TFs have been mapped from
human TFs using the homologene R-package (c.f. ./data-raw/process.R).
}
\author{
Sascha Jung \email{sjung@cicbiogune.es}
}
\keyword{data}
