% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MOUSE_tfs}
\alias{MOUSE_tfs}
\title{Transcription factors (Mouse)}
\format{
Character vector of length 1636.
}
\usage{
MOUSE_tfs
}
\description{
Transcription factors (Mouse)
}
\details{
The vector of mouse TFs corresponds to the values in the "Symbol" column of the
MOUSE_tf.db dataset.
}
\author{
Sascha Jung \email{sjung@cicbiogune.es}
}
\keyword{data}
