% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainlib.R
\name{path.prob.sum}
\alias{path.prob.sum}
\title{Calculating significance of receptor-TF paths}
\usage{
path.prob.sum(subg, iTF, Receptors, prob.matrix, ncores = 4)
}
\arguments{
\item{subg}{An igraph graph object}

\item{iTF}{A vector of interface TF names}

\item{Receptors}{A vector of receptor names}

\item{prob.matrix}{A matrix of probabilities for the interaction between genes i (rows) to j (columns)}

\item{ncores}{The number of cores used (default: 4)}
}
\value{
Data frame of z-scores per path
}
\description{
Calculates the z-score for each path between receptors and downstream TFs. Uses
all paths as background
}
