% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainlib.R
\name{product_path_weight}
\alias{product_path_weight}
\title{Calculating path weights}
\usage{
product_path_weight(path, graph)
}
\arguments{
\item{path}{A path as defined by igraph}

\item{graph}{An igraph graph object}
}
\value{
Numerical path weight
}
\description{
Returns a weight for a given path calculated as the product of edge weigths
}
