% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxFuncs.R
\name{reproduceNetworks}
\alias{reproduceNetworks}
\title{Reconstructs Pancreas networks for 7 and 9 Wpc}
\usage{
reproduceNetworks(weeks = c(7, 9), out.path)
}
\arguments{
\item{weeks}{A vector containing 7 and/or 9 (as integers) or both to signify which networks to reconstruct.}

\item{out.path}{Path to a folder where the output should be stored.}
}
\value{
Creates folders in the output directory containing the final interactomes of the requested networks.
The function itself returns NULL.
}
\description{
Reconstructs Pancreas networks for 7 and 9 Wpc.
}
