% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercom_library_coexp.R
\name{scoringFun}
\alias{scoringFun}
\title{Scoring ligand-receptor interactions}
\usage{
scoringFun(
  data,
  tissue.R.TF,
  tissue.LR,
  LR,
  sig.cutoff = 0.9,
  z.score.cutoff = 0
)
}
\arguments{
\item{data}{A data frame of gene expression values}

\item{tissue.R.TF}{Data frame of significant receptor-TF associations.}

\item{tissue.LR}{Data frame of potential ligand-receptor interactions}

\item{LR}{Data frame of ALL ligand-receptor interactions}

\item{sig.cutoff}{Significance cutoff as a percentile (ranges from 0 (most permissive) to 1 (most strict)) (default: 0.9)}

\item{z.score.cutoff}{An optional z-score cutoff from bootstrapping (DEPRECATED; DO NOT CHANGE)}
}
\value{
Data frame significant ligand-receptor interactions
}
\description{
Calculates the probability of interaction for each potential ligand-receptor association.
}
