% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mainlib.R
\name{spcal_path_weight}
\alias{spcal_path_weight}
\title{Calculating compatibility scores}
\usage{
spcal_path_weight(s, t, g)
}
\arguments{
\item{s}{A source node name present in graph g}

\item{t}{A character vector of target node names present in graph g}

\item{g}{An igraph graph object}
}
\value{
Compatibility score
}
\description{
Calculates the compatibility between a source node and a list of target nodes depending
on their expression values
}
