% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercom_library_coexp.R
\name{up_down_tfs}
\alias{up_down_tfs}
\title{Classify non-interface TFs as up or down regulated}
\usage{
up_down_tfs(g, deg, non_interface_TFs)
}
\arguments{
\item{g}{Markov chain of a signaling interactome}

\item{deg}{Data frame of differentially expressed genes (one column only)}

\item{non_interface_TFs}{TFs that are not regulated by any active signaling pathway}
}
\value{
Vector of differentially expressed non-interface TFs
}
\description{
Returns all non-interface TFs as up or down regulated
}
