% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/intercom_library_coexp.R
\name{weight_probability}
\alias{weight_probability}
\title{Calculating the weight probability}
\usage{
weight_probability(x)
}
\arguments{
\item{x}{Vector of weights across a path in the graph}
}
\value{
Compatibility score
}
\description{
Calculates the actual compatibility score given a vector of weights
}
